motion_sensor_app/requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
dnspython==2.7.0
eventlet==0.40.2
Flask==3.1.1
Flask-SocketIO==5.5.1
fonttools==4.60.1
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.2.4
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
RPi.GPIO==0.7.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
simple-websocket==1.1.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
Werkzeug==3.1.3
wsproto==1.2.0
xgboost==3.0.5
zope.event==6.0
zope.interface==8.0.1
